+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 * Added ```ParseAndValidate``` function.
 
+### Changed
+* **BREAKING**: Renamed ```NewProxyRequest``` function to ```ProxyRequest```.
+
 ## v0.5.0 - 2021-10-21
 ### Added
 * Added ```ProxyResponseHTML``` and ```ProxyResponseJSON``` functions.